requirements.txt
fastmcp
python-toon
pyyaml